pythonSerialPortReading/pyproject.toml
[tool.poetry]
name = "pythonSerialPortReading"
version = "0.1.0"
description = ""
authors = ["Rafael de Oliveira Évora <122195163+evoradev@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
